A2DP: Various Fixes in BT device disconnect path
- Set the flag when disconnect connection state is
sent to BT Apps and reset it back when state machine
transition happens successfully so that meanwhile we
return check A2DP ready as false even in open state.

- If index is full at time of connection, make sure
we send connection status of connecting device back
to disconnected to BT App so that Settings UI is not
stuck forever in CONNECTING State of that device.

Change-Id: I05480d4e85350302162c62dcb1bc561fe9e1097a