Proper handling of remote start for various dual A2dp usecases
 - Do not send suspend to both the devices when remote issues
Start when other device is active and DUT triggers suspend for
remote Start after 3s of timer expiry.
 - Do not mark device received remote Start as active device
Unless streaming starts on it.
 - Do not update audio state change to Apps for non-active device
when Remote Start is received when other index is actively streaming.
 - Do not use VSC command exchange state for nonpronto devices
as DUT always send full set of VSCs to SOC to start streaming.
 - Do not reset start initiated state to Start VSCs when Stop
preceeds Start for other index for handoff cases.
 - Postpone sending Suspend/Stop till remote_started alarm fires to
avoid putting remote in bad state.
 - If handoff index has remote start pending suspend, cancel the
3s timer which triggers the same when handoff is triggered. And
as we do not update audio state on remote start for non active device,
audio state to app is updated when start on that index actually happens
due to handoff.
 - Update audio state to App when DUT triggered start is received
when remote initiated start is in progress.
 - check for tx started flag too when reconfig due to handoff is in
progress and remote start is received to reset reconfig flag.
 - When local start overrides remote started state, reset dual handoff
flag as well and mark the index as current playing.
 - Issue Start for split A2dp solution on clearing remote AV suspend
flag on receipt of RC PLAY command.

Change-Id: I2af57b17b062c9d55acb7fce71484fbbe907e8fe