AV: handle remote start in reconfig state.
In an incoming connection case, if there is a
race condition between RECONFIG triggered from DUT
and remote START, the start request was not handled
in reconfig state. This causes AVDTP layer to
move to STREAM state however BTA and BTIF are
in RCFG and OPEN state respectively. After reconfig
is completed, BTA requests for START again, however
the AVDTP layer couldn't handle it.
The fix handles remote START in reconfig state so that
the link will be suspended after 3sec if no streaming
is started within this time.

CRs-Fixed: 2194667

Change-Id: I16aa00d0389845b2489e557c06d06e8b423545a4